requirements.txt
streamlit
pillow
streamlit-authenticator
pyyaml
